# Conflicts:
#	customerapp/src/main/java/com/tokopedia/tkpd/home/adapter/DataFeedAdapter.java
#	customerapp/src/main/java/com/tokopedia/tkpd/home/adapter/GridLayoutProductAdapter.java
#	customerapp/src/main/java/com/tokopedia/tkpd/home/adapter/WishlistRecyclerViewAdapter.java
#	tkpdcore/src/main/java/com/tokopedia/core/ManageShopAddress.java
#	tkpdcore/src/main/java/com/tokopedia/core/PreviewProductImage.java
#	tkpdcore/src/main/java/com/tokopedia/core/home/adapter/HistoryProductRecyclerViewAdapter.java
#	tkpdcore/src/main/java/com/tokopedia/core/manage/people/address/adapter/ChooseAddressAdapter.java
#	tkpdcore/src/main/java/com/tokopedia/core/product/presenter/ProductDetailPresenterImpl.java
#	tkpdcore/src/main/java/com/tokopedia/core/shopinfo/ShopInfoActivity.java
#	tkpdcore/src/main/java/com/tokopedia/core/shopinfo/adapter/ProductLargeDelegate.java
#	tkpdcore/src/main/java/com/tokopedia/core/shopinfo/adapter/ProductListDelegate.java
#	tkpdcore/src/main/java/com/tokopedia/core/shopinfo/adapter/ProductMediumDelegate.java
#	tkpdcore/src/main/java/com/tokopedia/core/util/MethodChecker.java
#	tkpdseller/src/main/java/com/tokopedia/seller/fragment/FragmentShopNewOrderDetailV2.java
#	tkpdseller/src/main/java/com/tokopedia/seller/fragment/FragmentShopShippingDetailV2.java
#	tkpdseller/src/main/java/com/tokopedia/seller/selling/view/activity/ActivitySellingTransaction.java

+++ tkpdcore/src/main/java/com/tokopedia/core/shopinfo/ReputationViewShop.java
@@ -635,15 +635,14 @@ public class ReputationViewShop extends TActivity {
         return new PopupMenu.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
-                switch (item.getItemId()) {
-                    case R2.id.action_report:
-                        showDialogReport();
-                        return true;
-                    case R2.id.action_delete:
-                        deleteComment();
-                        return true;
-                    default:
-                        return false;
+                if (item.getItemId() == R.id.action_report) {
+                    showDialogReport();
+                    return true;
+                } else if (item.getItemId() == R.id.action_delete) {
+                    deleteComment();
+                    return true;
+                } else {
+                    return false;
                 }
             }
         };
